refactor: Added features directory layer
