+++ ui/lib/features/presentation/components/feedback_snack_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:hours_control/presentation/themes/main_color_theme.dart';
+import 'package:hours_control/features/presentation/themes/main_color_theme.dart';
 
 enum SnackbarType {
   success,
