+++ ui/lib/features/presentation/screens/squads_screen.dart
@@ -2,52 +2,62 @@ import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:get_it/get_it.dart';
 import 'package:hours_control/core/mobx/platform_store.dart';
-import 'package:hours_control/presentation/components/action_button.dart';
-import 'package:hours_control/presentation/components/empty_data.dart';
-import 'package:hours_control/presentation/screens/dialogs/create_employee_dialog.dart';
-import 'package:hours_control/presentation/screens/dialogs/create_squad_dialog.dart';
-import 'package:hours_control/presentation/themes/main_color_theme.dart';
+import 'package:hours_control/features/presentation/components/action_button.dart';
+import 'package:hours_control/features/presentation/components/empty_data.dart';
+import 'package:hours_control/features/presentation/screens/dialogs/create_squad_dialog.dart';
+import 'package:hours_control/features/presentation/themes/main_color_theme.dart';
 
 final platformStore = GetIt.I.get<PlatformStore>();
 
-class EmployeesScreen extends StatefulWidget {
-  const EmployeesScreen({super.key});
+class SquadsScreen extends StatefulWidget {
+  const SquadsScreen({super.key});
 
   @override
-  State<EmployeesScreen> createState() => _EmployeesScreenState();
+  State<SquadsScreen> createState() => _SquadsScreenState();
 }
 
-class _EmployeesScreenState extends State<EmployeesScreen> {
+class _SquadsScreenState extends State<SquadsScreen> {
   late ScrollController _scrollController;
 
-  Future<void> _fetchEmployees() async {
+  Future<void> _fetchSquads() async {
     try {
-      platformStore.setIsFetchingEmployees(true);
+      platformStore.setIsFetchingSquads(true);
       await Future.delayed(const Duration(seconds: 2));
       // TODO: FETCH SQUADS
     } catch (e) {
-      print('Error fetching employees: $e');
+      print('Error fetching squads: $e');
     } finally {
-      platformStore.setIsFetchingEmployees(false);
+      platformStore.setIsFetchingSquads(false);
     }
   }
 
   List<DataRow> _buildTableData() {
     List<DataRow> dataRowList = [];
 
-    for (int i = 0; i < platformStore.employeeList.length; i++) {
-      final employee = platformStore.employeeList[i];
+    for (int i = 0; i < platformStore.squadList.length; i++) {
+      final squad = platformStore.squadList[i];
       dataRowList.add(
         DataRow(
           cells: <DataCell>[
             DataCell(
-              Text(employee.name),
+              Text(squad.id),
             ),
             DataCell(
-              Text(employee.estimatedHours),
+              Text(squad.name),
             ),
             DataCell(
-              Text(employee.squadId),
+              platformStore.isMobile
+                  ? IconButton(
+                      onPressed: () {},
+                      color: Theme.of(context).extension<MainColorTheme>()?.blue ?? Colors.blue,
+                      icon: const Icon(
+                        Icons.group,
+                        color: Colors.white,
+                      ),
+                    )
+                  : const ActionButton(
+                      text: "Visitar squad",
+                    ),
             ),
           ],
         ),
@@ -57,40 +67,12 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
     return [];
   }
 
-  String _getEmptyText() {
-    if (platformStore.squadList.isEmpty) {
-      return "Nenhuma squad cadastrada. Crie uma squad para começar.";
-    }
-    if (platformStore.employeeList.isEmpty) {
-      return "Nenhum usuário cadastrado. Crie um usuário para começar.";
-    }
-
-    return "";
-  }
-
-  String _getButtonText() {
-    if (platformStore.squadList.isEmpty) {
-      return "Criar squad";
-    }
-
-    return "Criar usuário";
-  }
-
-  Widget _getDialogContent() {
-    // TODO: Implement dialog content based on platformStore.squadList.isEmpty
-    // if (platformStore.squadList.isEmpty) {
-    //   return const CreateSquadDialog();
-    // }
-
-    return const CreateEmployeeDialog();
-  }
-
   @override
   void initState() {
     super.initState();
     _scrollController = ScrollController(initialScrollOffset: 0);
-    if (platformStore.squadList.isNotEmpty && platformStore.employeeList.isEmpty) {
-      _fetchEmployees();
+    if (platformStore.squadList.isEmpty) {
+      _fetchSquads();
     }
   }
 
@@ -107,7 +89,7 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Text(
-                "Lista de Usuários",
+                "Lista de Squads",
                 style: platformStore.isMobile
                     ? Theme.of(context).textTheme.headlineSmall
                     : Theme.of(context).textTheme.headlineMedium,
@@ -123,7 +105,7 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
                       ? const EdgeInsets.symmetric(horizontal: 8, vertical: 10)
                       : const EdgeInsets.symmetric(horizontal: 32, vertical: 64),
                   width: MediaQuery.of(context).size.width,
-                  child: platformStore.isFetchingEmployees
+                  child: platformStore.isFetchingSquads
                       ? Center(
                           child: CircularProgressIndicator(
                             color:
@@ -136,11 +118,11 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
                           children: [
                             Expanded(
                               flex: 1,
-                              child: platformStore.squadList.isEmpty ||
-                                      platformStore.employeeList.isEmpty
-                                  ? Center(
+                              child: platformStore.squadList.isEmpty
+                                  ? const Center(
                                       child: EmptyData(
-                                        emptyText: _getEmptyText(),
+                                        emptyText:
+                                            "Nenhuma squad cadastrada. Crie uma squad para começar.",
                                       ),
                                     )
                                   : SingleChildScrollView(
@@ -172,6 +154,14 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
                                               headingRowAlignment: MainAxisAlignment.start,
                                               label: Expanded(
                                                 child: Text(
+                                                  'ID',
+                                                ),
+                                              ),
+                                            ),
+                                            DataColumn(
+                                              headingRowAlignment: MainAxisAlignment.start,
+                                              label: Expanded(
+                                                child: Text(
                                                   'Nome',
                                                 ),
                                               ),
@@ -180,15 +170,7 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
                                               headingRowAlignment: MainAxisAlignment.start,
                                               label: Expanded(
                                                 child: Text(
-                                                  'Horas',
-                                                ),
-                                              ),
-                                            ),
-                                            DataColumn(
-                                              headingRowAlignment: MainAxisAlignment.start,
-                                              label: Expanded(
-                                                child: Text(
-                                                  'Squad ID',
+                                                  '',
                                                 ),
                                               ),
                                             ),
@@ -199,12 +181,12 @@ class _EmployeesScreenState extends State<EmployeesScreen> {
                                     ),
                             ),
                             ActionButton(
-                              text: _getButtonText(),
+                              text: "Criar squad",
                               onPressed: () {
                                 showDialog(
                                   context: context,
                                   builder: (BuildContext context) {
-                                    return _getDialogContent();
+                                    return const CreateSquadDialog();
                                   },
                                 );
                               },
